[EME] fix build after r195410

+++ Source/WebCore/platform/graphics/gstreamer/CDMPRSessionGStreamer.cpp
@@ -66,14 +66,14 @@ const String& CDMPRSessionGStreamer::sessionId() const
     return m_sessionId;
 }
 
-RefPtr<Uint8Array> CDMPRSessionGStreamer::generateKeyRequest(const String& mimeType, Uint8Array* initData, String& destinationURL, unsigned short& errorCode, unsigned long& systemCode)
+RefPtr<Uint8Array> CDMPRSessionGStreamer::generateKeyRequest(const String& mimeType, Uint8Array* initData, String& destinationURL, unsigned short& errorCode, uint32_t& systemCode)
 {
     UNUSED_PARAM(mimeType);
 
     return dxdrmGenerateKeyRequest(initData, destinationURL, errorCode, systemCode);
 }
 
-bool CDMPRSessionGStreamer::update(Uint8Array* key, RefPtr<Uint8Array>& nextMessage, unsigned short& errorCode, unsigned long& systemCode)
+bool CDMPRSessionGStreamer::update(Uint8Array* key, RefPtr<Uint8Array>& nextMessage, unsigned short& errorCode, uint32_t& systemCode)
 {
     return dxdrmProcessKey(key, nextMessage, errorCode, systemCode);
 }
